requirements.txt
altair==4.0.1
altair_saver==0.5.0
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.3
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.18.5
oauthlib==3.1.0
pandas==1.0.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sendgrid==6.3.1
sendgrid-python==0.1.1
six==1.15.0
soupsieve==2.0.1
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
